Extracting loglist testing data in a proper way.
